requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.4
billiard==3.6.3.0
boto==2.49.0
boto3==1.13.24
botocore==1.16.24
CacheControl==0.12.6
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.5
django-amazon-ses==3.0.2
django-cors-headers==3.2.1
django-filter==2.4.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
drf-yasg==1.17.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
ipaddr==2.2.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.8
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.2.0
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
Pillow==7.1.1
pip-autoremove==0.9.1
pluggy==0.13.1
progress==1.5
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.8.1
pycodestyle==2.6.0
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-slugify==4.0.0
pytoml==0.1.21
pytz==2019.3
redis==3.5.0
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
